# docs/swagger.yaml
basePath: /
definitions:
  api.CardItemResponse:
    properties:
      checksum:
        type: string
      created_at:
        type: string
      holder:
        type: string
      metadata:
        type: string
      name:
        type: string
      number:
        type: string
      updated_at:
        type: string
    type: object
  api.CardResponse:
    properties:
      checksum:
        type: string
      created_at:
        type: string
      cvv:
        type: string
      expires_at:
        type: string
      holder:
        type: string
      metadata:
        type: string
      name:
        type: string
      number:
        type: string
      updated_at:
        type: string
    type: object
  api.CreateCardRequest:
    properties:
      cvv:
        type: string
      expires_at:
        type: string
      holder:
        type: string
      metadata:
        type: string
      name:
        type: string
      number:
        type: string
    type: object
...
  api.UpdateCardRequest:
    properties:
      cvv:
        type: string
      expires_at:
        type: string
      holder:
        type: string
      metadata:
        type: string
      number:
        type: string
    type: object
...
info:
  contact: {}
  description: Менеджер паролей GophKeeper
  title: GophKeeper API
  version: "1.0"
paths:
  /api/card:
    get:
      description: Возвращает карты текущего пользователя без CVV и срока действия
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.CardItemResponse'
            type: array
        "401":
          description: Пользователь не аутентифицирован
          schema:
            type: string
        "500":
          description: Внутренняя ошибка
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Список карт пользователя
      tags:
      - card
    post:
      consumes:
      - application/json
      description: Сохраняет данные банковской карты текущего пользователя. Имя карты
        уникально в рамках пользователя
      parameters:
      - description: Данные карты
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.CreateCardRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.CardResponse'
        "400":
          description: Неверный формат запроса
          schema:
            type: string
        "401":
          description: Пользователь не аутентифицирован
          schema:
            type: string
        "409":
          description: Карта с таким именем уже существует
          schema:
            type: string
        "500":
          description: Внутренняя ошибка
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Создание банковской карты
      tags:
      - card
  /api/card/{card_name}:
    delete:
      description: Удаляет карту текущего пользователя по имени
      parameters:
      - description: Имя карты
        in: path
        name: card_name
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "204":
          description: No Content
        "400":
          description: Неверный формат запроса
          schema:
            type: string
        "401":
          description: Пользователь не аутентифицирован
          schema:
            type: string
        "404":
          description: Карта не найдена
          schema:
            type: string
        "500":
          description: Внутренняя ошибка
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Удаление карты
      tags:
      - card
    get:
      description: Возвращает данные карты текущего пользователя по имени
      parameters:
      - description: Имя карты
        in: path
        name: card_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.CardResponse'
        "400":
          description: Неверный формат запроса
          schema:
            type: string
        "401":
          description: Пользователь не аутентифицирован
          schema:
            type: string
        "404":
          description: Карта не найдена
          schema:
            type: string
        "500":
          description: Внутренняя ошибка
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Получение карты
      tags:
      - card
    put:
      consumes:
      - application/json
      description: Полностью перезаписывает данные карты текущего пользователя, найденной
        по имени
      parameters:
      - description: Имя карты
        in: path
        name: card_name
        required: true
        type: string
      - description: Новые данные карты
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.UpdateCardRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.CardResponse'
        "400":
          description: Неверный формат запроса
          schema:
            type: string
        "401":
          description: Пользователь не аутентифицирован
          schema:
            type: string
        "404":
          description: Карта не найдена
          schema:
            type: string
        "500":
          description: Внутренняя ошибка
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Обновление карты
      tags:
      - card
...
securityDefinitions:
  BearerAuth:
    description: 'Введите токен в формате: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"